Properly consolidate error handling by modifying existing classes (Issue #189)
Integrate error handling directly into core CRUDArticle class instead of creating parallel implementations:

- Update CRUDArticle to inherit from ErrorHandledCRUD instead of CRUDBase
- Add factory pattern for DI compatibility
- Update tests to expect error raising behavior instead of returning None
- Remove redundant error_handled_article.py implementation
